+++ src/common/components/alert/alert-toast.styles.tsx
@@ -7,22 +7,22 @@ export const AlertToast = styled(Alert)<{isSmall: boolean}>`
 `;
 
 export const Toast = styled.div<{isSmall: boolean}>`
-  width: ${props => props.isSmall ? '290px' : '350px'};
-  height: 60px;
-  border-top: 4px solid ${props => props.theme.suomifi.colors.successBase};
-  border-radius: 3px 3px 0px 0px;
-  background: ${props => props.theme.suomifi.colors.whiteBase};
-  display: flex;
   align-items: center;
+  background: ${props => props.theme.suomifi.colors.whiteBase};
+  border-radius: 3px;
+  border-top: 4px solid ${props => props.theme.suomifi.colors.successBase};
+  display: flex;
   font-size: 16px;
   gap: 10px;
+  height: 60px;
+  width: ${props => props.isSmall ? '290px' : '350px'};
 `;
 
 export const ToastIcon = styled(Icon)`
-  height: 24px;
-  width: 24px;
-  margin-left: 15px;
   color: ${props => props.theme.suomifi.colors.successBase};
+  height: 24px;
+  margin-left: 15px;
+  width: 24px;
 `;
 
 export const AlertsWrapper = styled.div`
